+++ packages/std/src/math/decimal256.rs
@@ -4,7 +4,7 @@ use serde::{de, ser, Deserialize, Deserializer, Serialize};
 use std::cmp::Ordering;
 use std::convert::TryInto;
 use std::fmt::{self, Write};
-use std::ops::{Add, AddAssign, Div, DivAssign, Mul, Sub};
+use std::ops::{Add, AddAssign, Div, DivAssign, Mul, Sub, SubAssign};
 use std::str::FromStr;
 use thiserror::Error;
 
@@ -362,6 +362,14 @@ impl Sub for Decimal256 {
         Self(self.0 - other.0)
     }
 }
+forward_ref_binop!(impl Sub, sub for Decimal256, Decimal256);
+
+impl SubAssign for Decimal256 {
+    fn sub_assign(&mut self, rhs: Decimal256) {
+        *self = *self - rhs;
+    }
+}
+forward_ref_op_assign!(impl SubAssign, sub_assign for Decimal256, Decimal256);
 
 impl Mul for Decimal256 {
     type Output = Self;